서버 코드 분리
