+++ server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "jest",
     "start": "nodemon --exec node --loader ts-node/esm src/app.js",
-    "watch": "watchify src/public/index.js -o src/public/bundle.js -v"
+    "watch": "watchify public/index.js -o public/bundle.js -v"
   },
   "author": "amir eshaq",
   "license": "ISC",
